PDDA-474 Moving getCourtSites() out of DisplayService into AbstractService
